+++ attribution/src/main/java/com/affise/attribution/errors/AffiseError.kt
@@ -1,9 +1,11 @@
 package com.affise.attribution.errors
 
 sealed class AffiseError(message: String) : Exception(message) {
-    class AlreadyInitialized: AffiseError(message = MESSAGE_ALREADY_INITIALIZED)
+    class AlreadyInitialized: AffiseError(message = ALREADY_INITIALIZED)
 
     companion object {
-        internal const val MESSAGE_ALREADY_INITIALIZED = "Affise SDK is already initialized"
+        const val ALREADY_INITIALIZED = "Affise SDK is already initialized"
+        const val NOT_INITIALIZED = "Affise SDK is not initialized"
+        const val ERROR_READING_FROM_PREFERENCES = "error_reading_from_preferences"
     }
 }